# Addressing performance issues with broadcasting