This PR introduces a "lazy" version of `arraydist` which allows us to hack around performance issues with broadcasting over constructors in ReverseDiff and Zygote. Based on "insights" from https://github.com/TuringLang/Turing.jl/issues/1934.

More specifically, in broadcasting:
- Zygote uses ForwardDiff if certain checks are satisfied here: https://github.com/FluxML/Zygote.jl/blob/c2f1794ca9da3088a2f3bfb0144c8bfc4dd89d9a/src/lib/broadcast.jl#L198-L199. Having a type in the broadcasting statement means the combined eltype (https://github.com/FluxML/Zygote.jl/blob/c2f1794ca9da3088a2f3bfb0144c8bfc4dd89d9a/src/lib/broadcast.jl#L194) isn't going to be a subtype of `Union{Real,Complex}`, hence we miss the fast branch.
- ReverseDiff always uses ForwardDiff, but is very prone to type-instabilities when constructors are involved in the broadcast statement. One way to fix this is to just use a function instead of a type in the broadcast statement, e.g. replace `Normal` with `NormalF(args...) = Normal(args...)` in a broadcast will actually help ReverseDiff (https://github.com/TuringLang/Turing.jl/issues/1934#issuecomment-1382244113).

Here's an example of the result of the "lazy" array dist:

```julia
julia> using Distributions, DistributionsAD, LogDensityProblems, LogDensityProblemsAD, ReverseDiff, BenchmarkTools

julia> # Examples.
       n = 1000;

julia> logitp = randn(n); x = rand(Bool, n);

julia> struct Slow{A}
           x::A
       end

julia> LogDensityProblems.logdensity(f::Slow, logitp) = sum(logpdf.(BernoulliLogit.(logitp), f.x))

julia> LogDensityProblems.dimension(f::Slow) = length(f.x)

julia> LogDensityProblems.capabilities(::Type{<:Slow}) = LogDensityProblems.LogDensityOrder{0}()

julia> struct Fast{A}
           x::A
       end

julia> LogDensityProblems.logdensity(f::Fast, logitp) = logpdf(arraydist(BernoulliLogit, logitp), f.x)

julia> LogDensityProblems.dimension(f::Fast) = length(f.x)

julia> LogDensityProblems.capabilities(::Type{<:Fast}) = LogDensityProblems.LogDensityOrder{0}()

julia> adbackend = :ReverseDiff
:ReverseDiff

julia> kwargs = if adbackend === :ReverseDiff
           (compile=Val(true),)
       else
           NamedTuple()
       end
(compile = Val{true}(),)

julia> ∂slow = ADgradient(adbackend, Slow(x); kwargs...)
ReverseDiff AD wrapper for Slow{Vector{Bool}}(Bool[0, 1, 1, 0, 0, 1, 1, 1, 1, 0  …  0, 0, 0, 0, 1, 1, 0, 0, 1, 1]) (compiled tape)

julia> ∂fast = ADgradient(adbackend, Fast(x); kwargs...)
ReverseDiff AD wrapper for Fast{Vector{Bool}}(Bool[0, 1, 1, 0, 0, 1, 1, 1, 1, 0  …  0, 0, 0, 0, 1, 1, 0, 0, 1, 1]) (compiled tape)

julia> @benchmark $(LogDensityProblems.logdensity_and_gradient)($∂slow, $logitp)
BenchmarkTools.Trial: 5650 samples with 1 evaluation.
 Range (min … max):  838.543 μs …   4.950 ms  ┊ GC (min … max): 0.00% … 80.79%
 Time  (median):     847.915 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   883.237 μs ± 306.428 μs  ┊ GC (mean ± σ):  2.82% ±  6.62%

  ▄▇█▆▅▄▃▃▂▂▂   ▁▂▂▁▁                                           ▁
  ████████████▇████████▇▆▆▅▅▆▃▆▅▅▆███▇▆▄▄▆▅▅▄▄▄▅▄▆▅▃▃▁▅▅▃▅▄▃▃▄▃ █
  839 μs        Histogram: log(frequency) by time       1.05 ms <

 Memory estimate: 406.23 KiB, allocs estimate: 13491.

julia> @benchmark $(LogDensityProblems.logdensity_and_gradient)($∂fast, $logitp)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  32.306 μs …  2.524 ms  ┊ GC (min … max): 0.00% … 98.21%
 Time  (median):     34.354 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   34.874 μs ± 25.074 μs  ┊ GC (mean ± σ):  0.71% ±  0.98%

  ▄█▆▂   ▃    ▁██▆▄▃▂▅▆▄▁▁ ▁▂▁     ▂      ▂▃       ▂▁▁        ▂
  █████▄▆██▆▅▅████████████████▇▇▆▅▆██▆▆▅▃▃███▇▆▆▅▆▆████▇▇▆▆▅▆ █
  32.3 μs      Histogram: log(frequency) by time      40.8 μs <

 Memory estimate: 7.97 KiB, allocs estimate: 2.

julia> using Zygote

julia> adbackend = :Zygote
:Zygote

julia> kwargs = if adbackend === :ReverseDiff
           (compile=Val(true),)
       else
           NamedTuple()
       end
NamedTuple()

julia> ∂slow = ADgradient(adbackend, Slow(x); kwargs...)
Zygote AD wrapper for Slow{Vector{Bool}}(Bool[0, 1, 1, 0, 0, 1, 1, 1, 1, 0  …  0, 0, 0, 0, 1, 1, 0, 0, 1, 1])

julia> ∂fast = ADgradient(adbackend, Fast(x); kwargs...)
Zygote AD wrapper for Fast{Vector{Bool}}(Bool[0, 1, 1, 0, 0, 1, 1, 1, 1, 0  …  0, 0, 0, 0, 1, 1, 0, 0, 1, 1])

julia> @benchmark $(LogDensityProblems.logdensity_and_gradient)($∂slow, $logitp)
BenchmarkTools.Trial: 4265 samples with 1 evaluation.
 Range (min … max):  1.076 ms …  10.724 ms  ┊ GC (min … max): 0.00% … 88.85%
 Time  (median):     1.106 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.170 ms ± 652.314 μs  ┊ GC (mean ± σ):  4.76% ±  7.43%

   █                                                           
  ▅██▇▆▄▅▆▆▆▆▅▄▄▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▂▂▂▂▁▁▂▂▂▂ ▃
  1.08 ms         Histogram: frequency by time        1.37 ms <

 Memory estimate: 536.72 KiB, allocs estimate: 14601.

julia> @benchmark $(LogDensityProblems.logdensity_and_gradient)($∂fast, $logitp)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  189.582 μs …   8.207 ms  ┊ GC (min … max): 0.00% … 96.49%
 Time  (median):     195.637 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   206.669 μs ± 194.132 μs  ┊ GC (mean ± σ):  2.77% ±  2.88%

   ▄▇█▇▄▄▆▆▃▂▁▂▃▅▅▄▃▂▂▂▂▁▁    ▁▁                                ▂
  ▆████████████████████████▇▇████▇▇███▇▇▇▇▇▇█▆▆▇▇▇▆▆▇▆▆▆▅▅▅▅▅▅▅ █
  190 μs        Histogram: log(frequency) by time        257 μs <

 Memory estimate: 68.69 KiB, allocs estimate: 473.
```

I also thought maybe this is where LazyArrays.jl could be useful, but preliminary attempts didn't end up being fruitful so uncertain.