@torfjelde 2023-01-15
I am of course in complete agreement with you in everything you say here, but do you think such a "hacky" approach would be accepted into Distributions?
It's seems to be an issue with either AD implantation or the Julia type inference that should really be addressed, but that is unfortunately beyond my own capability and will probably require a lot of work.

That is, do you foresee another solution in the near- or even medium-term future?:/