@devmotion 2023-01-15
Could we add performance and AD improvements to `ProductDistribution`s with `FillArray`s and broadcast objects resulting from `Broadcast.instantiate(Broadcast.broadcasted(....))` to Distributions? The main problem with the latter might be to infer the type of the distribution correctly without materializing everything.