@torfjelde 2023-01-15
> Could we add performance and AD improvements to ProductDistributions with FillArrays and broadcast objects resulting from Broadcast.instantiate(Broadcast.broadcasted(....)) to Distributions?

I'm not completlely understanding what you mean here :confused: Do you mean we should make something like

```julia
product_distribution(Broadcast.broadcasted(BernoulliLogit, logitp))
```

work and have a similar implementation as in this PR?