@torfjelde 2023-01-15
Maybe something like (I know `Product` is deprecated but it's a bit simpler than `ProductDistribution`):

``` julia
using Distributions: UnivariateDistribution, MultivariateDistribution, ValueSupport

struct LazyProduct{
    S<:ValueSupport,
    T<:UnivariateDistribution{S},
    V,
} <: MultivariateDistribution{S}
    v::V
    function LazyProduct{S,T,V}(v::V) where {S<:ValueSupport,T<:UnivariateDistribution{S},V}
        return new{S,T,V}(v)
    end
end

function LazyProduct(
    v::V
) where {S<:ValueSupport,T<:UnivariateDistribution{S},V<:Broadcast.Broadcasted{<:Any,<:Any,Type{T}}}
    return LazyProduct{S, T, V}(v)
end

Base.length(d::LazyProduct) = length(d.v)
function Base.eltype(::Type{<:LazyProduct{S,T}}) where {S<:ValueSupport,
                                                        T<:UnivariateDistribution{S}}
    return eltype(Broadcast.combine_eltypes(bc.f, bc.args))
end

Distributions._rand!(rng::Distributions.AbstractRNG, d::LazyProduct, x::AbstractVector{<:Real}) =
    map!(Base.Fix1(rand, rng), x, d.v)

function Distributions._logpdf(d::LazyProduct, x::AbstractVector{<:Real})
    bc = d.v
    f = make_logpdf_closure(bc.f)
    return sum(f.(x, bc.args...))
end
```

This does the trick for ReverseDiff but Zygote complains (it tries to call `LazyDistribution(::AbstractVector)` at some point, so might just require a custom rule).