@devmotion 2023-01-15
Basically yes but for `Broadcast.instantiate(Broadcast.broadcasted(...))`. Similar to how base supports it in eg `mapreduce`, `sum` etc. (often eg `sum(Broadcast.instantiate(Broadcast.broadcasted(f, x, y)))` is a very performant alternative to `sum(f(xi, yi) for (xi, yi) in zip(x, y))` and `sum(f.(x, y))` or `sum(map(f, x, y))` that uses pairwise summation (in contrast to `zip`) and avoids materializing the array (in contrast to `broadcast` and `map`) - therefore it's used e.g. in Distributions and StatsBase IIRC).