@torfjelde 2023-01-15
> sum(Broadcast.instantiate(Broadcast.broadcasted(f, x, y)))

This is much slower for ReverseDiff vs. actually allocating, i.e. `sum(f.(x, y))` :confused: Maybe this could be fixed in ReverseDiff though..