@devmotion 2023-01-15
IMO `Broadcast.instantiate(...)` is so common (e.g. it's also used by LazyArrays: https://github.com/JuliaArrays/LazyArrays.jl/blob/09f382f8d4828d37eb11474963b671dfa95ea43d/src/lazymacro.jl#L114) that it's a bug if an AD system does not support it and these bugs should be fixed.