@torfjelde 2023-01-15
Fair :confused: 

Speaking of LazyArrays; I was just looking at maybe using this for the product distribution. Then it just becomes a matter of doing `product_distribution(BroadcastArray(BernoulliLogit, logitp))`, etc.