@torfjelde 2023-01-15
> Broadcast.instantiate(...)

Just for the record, ReverseDiff supports this and all. It's just that `instantiate(broadcasted(...))` seems, at a first glance, to be slower than `f.(...)` directly.