@torfjelde 2023-01-15
Or maybe I did something wrong because now I can't reproduce it :upside_down_face: