@torfjelde 2023-01-15
That is, we can just change

https://github.com/TuringLang/DistributionsAD.jl/blob/c30bcd993e5a0f0cdb56336c5e4a18cb63f198f9/src/DistributionsAD.jl#L91-L96

to something like

```julia
        make_logpdf_closure(::Type{<:BroadcastVector{<:Any,Type{F}}}) where {F} = make_logpdf_closure(F)

        function Distributions._logpdf(
            dist::LazyVectorOfUnivariate,
            x::AbstractVector{<:Real},
        )
            f = DistributionsAD.make_logpdf_closure(typeof(dist.v))
            # TODO: Fix ReverseDiff performance on this.
            return sum(Broadcast.instantiate(Broadcast.broadcasted(f, x, dist.v.args...)))
        end
```

with

```julia
make_logpdf_closure(::Type{D}) where {D} = (x, args...) -> logpdf(D(args...), x)
```