@torfjelde 2023-01-15
Regarding ReverseDiff's issue with `sum(instantiate(broadcasted(f, ...)))` vs. `sum(f.(...))` is that the former traces through the _entire_ computation using `TrackedReal` (resulting in a long tape) while the latter results in a single `∇broadcast` and a `sum`, i.e. the tape is only of length 2 (and we end up using ForwardDiff for `∇broadcast`)