@torfjelde 2023-01-15
> Regarding ReverseDiff's issue with sum(instantiate(broadcasted(f, ...))) vs. sum(f.(...)) is that the former traces through the entire computation using TrackedReal (resulting in a long tape) while the latter results in a single ∇broadcast and a sum, i.e. the tape is only of length 2 (and we end up using ForwardDiff for ∇broadcast)

This seems like it's somewhat non-trivial to address :confused: The issue is that we _want_ to allocate in the case we're using ReverseDiff since it allows us to vectorize using ForwardDiff..